# bcm283x: boost Pin.Out perf by 25%