Timings for a full cycle (High, Low):
- RPi3:  was 610ns 1.6Mhz, now 460ns, 2.2Mhz (25%)
- RPiZW: was 585ns 1.7Mhz, now 475ns, 2.1Mhz (19%)

It seems that the boundary offset took an unreasonable amount of time compared
to the rest of the execution. Using hard coded offsets in the array with
conditions proved to be significantly faster.